+++ src/main/java/com/beniregev/custom_annotation_example/pizza/PizzaDao.java
@@ -8,7 +8,7 @@ public interface PizzaDao {
 
 	List<PizzaMenu> getPizzaMenus();
 
-	List<PizzaMenu> getPizzaMenus(String menuName, Integer minPrice, Integer maxPrice);
+	List<PizzaMenu> getPizzaMenus(String menuName, Double minPrice, Double maxPrice);
 
 	PizzaMenu getPizza(Long id);
 }
